Move CheckpointSet transform to internal package
